+++ atlas-content-modeler.php
@@ -56,7 +56,9 @@ function atlas_content_modeler_loader(): void {
 		'blueprints/export.php',
 		'api/crud-functions.php',
 		'api/validation-functions.php',
+		'api/utility-functions.php',
 		'class-validation-exception.php',
+		'class-wp-error.php',
 	);
 
 	foreach ( $plugin_files as $file ) {